+++ libsolidity/formal/CVC4Interface.cpp
@@ -48,9 +48,10 @@ void CVC4Interface::pop()
 	m_solver.pop();
 }
 
-void CVC4Interface::declareVariable(string const& _name, Sort const& _sort)
+void CVC4Interface::declareVariable(string const& _name, SortPointer const& _sort)
 {
-	m_variables[_name] = m_context.mkVar(_name.c_str(), cvc4Sort(_sort));
+	solAssert(_sort, "");
+	m_variables[_name] = m_context.mkVar(_name.c_str(), cvc4Sort(*_sort));
 }
 
 void CVC4Interface::addAssertion(Expression const& _expr)