+++ libsolidity/formal/ModelChecker.cpp
@@ -24,7 +24,7 @@ using namespace dev::solidity;
 
 ModelChecker::ModelChecker(ErrorReporter& _errorReporter, map<h256, string> const& _smtlib2Responses):
 	m_bmc(m_context, _errorReporter, _smtlib2Responses),
-	m_chc(m_context, _errorReporter),
+	m_chc(m_context, _errorReporter, _smtlib2Responses),
 	m_context()
 {
 }
@@ -40,5 +40,5 @@ void ModelChecker::analyze(SourceUnit const& _source)
 
 vector<string> ModelChecker::unhandledQueries()
 {
-	return m_bmc.unhandledQueries();
+	return m_bmc.unhandledQueries() + m_chc.unhandledQueries();
 }