+++ test/evmc/evmc.hpp
@@ -7,6 +7,7 @@
 #include <evmc/evmc.h>
 #include <evmc/helpers.h>
 
+#include <functional>
 #include <initializer_list>
 #include <utility>
 
@@ -14,6 +15,250 @@
 /// @ingroup cpp
 namespace evmc
 {
+/// The big-endian 160-bit hash suitable for keeping an Ethereum address.
+///
+/// This type wraps C ::evmc_address to make sure objects of this type are always initialized.
+struct address : evmc_address
+{
+    /// Default and converting constructor.
+    ///
+    /// Initializes bytes to zeros if not other @p init value provided.
+    constexpr address(evmc_address init = {}) noexcept : evmc_address{init} {}
+
+    /// Explicit operator converting to bool.
+    constexpr inline explicit operator bool() const noexcept;
+};
+
+/// The fixed size array of 32 bytes for storing 256-bit EVM values.
+///
+/// This type wraps C ::evmc_bytes32 to make sure objects of this type are always initialized.
+struct bytes32 : evmc_bytes32
+{
+    /// Default and converting constructor.
+    ///
+    /// Initializes bytes to zeros if not other @p init value provided.
+    constexpr bytes32(evmc_bytes32 init = {}) noexcept : evmc_bytes32{init} {}
+
+    /// Explicit operator converting to bool.
+    constexpr inline explicit operator bool() const noexcept;
+};
+
+/// The alias for evmc::bytes32 to represent a big-endian 256-bit integer.
+using uint256be = bytes32;
+
+
+/// Loads 64 bits / 8 bytes of data from the given @p bytes array in big-endian order.
+constexpr inline uint64_t load64be(const uint8_t* bytes) noexcept
+{
+    // TODO: Report bug in clang incorrectly optimizing this with AVX2 enabled.
+    return (uint64_t{bytes[0]} << 56) | (uint64_t{bytes[1]} << 48) | (uint64_t{bytes[2]} << 40) |
+           (uint64_t{bytes[3]} << 32) | (uint64_t{bytes[4]} << 24) | (uint64_t{bytes[5]} << 16) |
+           (uint64_t{bytes[6]} << 8) | uint64_t{bytes[7]};
+}
+
+/// Loads 32 bits / 4 bytes of data from the given @p bytes array in big-endian order.
+constexpr inline uint32_t load32be(const uint8_t* bytes) noexcept
+{
+    return (uint32_t{bytes[0]} << 24) | (uint32_t{bytes[1]} << 16) | (uint32_t{bytes[2]} << 8) |
+           uint32_t{bytes[3]};
+}
+
+namespace fnv
+{
+constexpr auto prime = 0x100000001b3;              ///< The 64-bit FNV prime number.
+constexpr auto offset_basis = 0xcbf29ce484222325;  ///< The 64-bit FNV offset basis.
+
+/// The hashing transformation for 64-bit inputs based on the FNV-1a formula.
+constexpr inline uint64_t fnv1a_by64(uint64_t h, uint64_t x) noexcept
+{
+    return (h ^ x) * prime;
+}
+}  // namespace fnv
+
+
+/// The "equal" comparison operator for the evmc::address type.
+constexpr bool operator==(const address& a, const address& b) noexcept
+{
+    // TODO: Report bug in clang keeping unnecessary bswap.
+    return load64be(&a.bytes[0]) == load64be(&b.bytes[0]) &&
+           load64be(&a.bytes[8]) == load64be(&b.bytes[8]) &&
+           load32be(&a.bytes[16]) == load32be(&b.bytes[16]);
+}
+
+/// The "not equal" comparison operator for the evmc::address type.
+constexpr bool operator!=(const address& a, const address& b) noexcept
+{
+    return !(a == b);
+}
+
+/// The "less" comparison operator for the evmc::address type.
+constexpr bool operator<(const address& a, const address& b) noexcept
+{
+    return load64be(&a.bytes[0]) < load64be(&b.bytes[0]) ||
+           (load64be(&a.bytes[0]) == load64be(&b.bytes[0]) &&
+            load64be(&a.bytes[8]) < load64be(&b.bytes[8])) ||
+           (load64be(&a.bytes[8]) == load64be(&b.bytes[8]) &&
+            load32be(&a.bytes[16]) < load32be(&b.bytes[16]));
+}
+
+/// The "equal" comparison operator for the evmc::bytes32 type.
+constexpr bool operator==(const bytes32& a, const bytes32& b) noexcept
+{
+    return load64be(&a.bytes[0]) == load64be(&b.bytes[0]) &&
+           load64be(&a.bytes[8]) == load64be(&b.bytes[8]) &&
+           load64be(&a.bytes[16]) == load64be(&b.bytes[16]) &&
+           load64be(&a.bytes[24]) == load64be(&b.bytes[24]);
+}
+
+/// The "not equal" comparison operator for the evmc::bytes32 type.
+constexpr bool operator!=(const bytes32& a, const bytes32& b) noexcept
+{
+    return !(a == b);
+}
+
+/// The "less" comparison operator for the evmc::bytes32 type.
+constexpr bool operator<(const bytes32& a, const bytes32& b) noexcept
+{
+    return load64be(&a.bytes[0]) < load64be(&b.bytes[0]) ||
+           (load64be(&a.bytes[0]) == load64be(&b.bytes[0]) &&
+            load64be(&a.bytes[8]) < load64be(&b.bytes[8])) ||
+           (load64be(&a.bytes[8]) == load64be(&b.bytes[8]) &&
+            load64be(&a.bytes[16]) < load64be(&b.bytes[16])) ||
+           (load64be(&a.bytes[16]) == load64be(&b.bytes[16]) &&
+            load64be(&a.bytes[24]) < load64be(&b.bytes[24]));
+}
+
+/// Checks if the given address is the zero address.
+constexpr inline bool is_zero(const address& a) noexcept
+{
+    return a == address{};
+}
+
+constexpr address::operator bool() const noexcept
+{
+    return !is_zero(*this);
+}
+
+/// Checks if the given bytes32 object has all zero bytes.
+constexpr inline bool is_zero(const bytes32& a) noexcept
+{
+    return a == bytes32{};
+}
+
+constexpr bytes32::operator bool() const noexcept
+{
+    return !is_zero(*this);
+}
+
+namespace literals
+{
+namespace internal
+{
+template <typename T, T... Ints>
+struct integer_sequence
+{
+};
+
+template <uint8_t... Bytes>
+using byte_sequence = integer_sequence<uint8_t, Bytes...>;
+
+template <char... Chars>
+using char_sequence = integer_sequence<char, Chars...>;
+
+
+template <typename, typename>
+struct concatenate;
+
+template <uint8_t... Bytes1, uint8_t... Bytes2>
+struct concatenate<byte_sequence<Bytes1...>, byte_sequence<Bytes2...>>
+{
+    using type = byte_sequence<Bytes1..., Bytes2...>;
+};
+
+template <uint8_t D>
+constexpr uint8_t parse_hex_digit() noexcept
+{
+    static_assert((D >= '0' && D <= '9') || (D >= 'a' && D <= 'f') || (D >= 'A' && D <= 'F'),
+                  "literal must be hexadecimal integer");
+    return static_cast<uint8_t>(
+        (D >= '0' && D <= '9') ? D - '0' : (D >= 'a' && D <= 'f') ? D - 'a' + 10 : D - 'A' + 10);
+}
+
+
+template <typename>
+struct parse_digits;
+
+template <uint8_t Digit1, uint8_t Digit2>
+struct parse_digits<byte_sequence<Digit1, Digit2>>
+{
+    using type = byte_sequence<static_cast<uint8_t>(parse_hex_digit<Digit1>() << 4) |
+                               parse_hex_digit<Digit2>()>;
+};
+
+template <uint8_t Digit1, uint8_t Digit2, uint8_t... Rest>
+struct parse_digits<byte_sequence<Digit1, Digit2, Rest...>>
+{
+    using type = typename concatenate<typename parse_digits<byte_sequence<Digit1, Digit2>>::type,
+                                      typename parse_digits<byte_sequence<Rest...>>::type>::type;
+};
+
+
+template <typename, typename>
+struct parse_literal;
+
+template <typename T, char Prefix1, char Prefix2, char... Literal>
+struct parse_literal<T, char_sequence<Prefix1, Prefix2, Literal...>>
+{
+    static_assert(Prefix1 == '0' && Prefix2 == 'x', "literal must be in hexadecimal notation");
+    static_assert(sizeof...(Literal) == sizeof(T) * 2, "literal must match the result type size");
+
+    template <uint8_t... Bytes>
+    static constexpr T create_from(byte_sequence<Bytes...>) noexcept
+    {
+        return T{{{Bytes...}}};
+    }
+
+    static constexpr T get() noexcept
+    {
+        return create_from(typename parse_digits<byte_sequence<Literal...>>::type{});
+    }
+};
+
+template <typename T, char Digit>
+struct parse_literal<T, char_sequence<Digit>>
+{
+    static_assert(Digit == '0', "only 0 is allowed as a single digit literal");
+    static constexpr T get() noexcept { return {}; }
+};
+
+template <typename T, char... Literal>
+constexpr T parse() noexcept
+{
+    return parse_literal<T, char_sequence<Literal...>>::get();
+}
+}  // namespace internal
+
+/// Literal for evmc::address.
+template <char... Literal>
+constexpr address operator"" _address() noexcept
+{
+    return internal::parse<address, Literal...>();
+}
+
+/// Literal for evmc::bytes32.
+template <char... Literal>
+constexpr bytes32 operator"" _bytes32() noexcept
+{
+    return internal::parse<bytes32, Literal...>();
+}
+}  // namespace literals
+
+using namespace literals;
+
+
+/// Alias for evmc_make_result().
+constexpr auto make_result = evmc_make_result;
+
 /// @copydoc evmc_result
 ///
 /// This is a RAII wrapper for evmc_result and objects of this type
@@ -26,6 +271,22 @@ public:
     using evmc_result::output_data;
     using evmc_result::output_size;
     using evmc_result::status_code;
+
+    /// Creates the result from the provided arguments.
+    ///
+    /// The provided output is copied to memory allocated with malloc()
+    /// and the evmc_result::release function is set to one invoking free().
+    ///
+    /// @param _status_code  The status code.
+    /// @param _gas_left     The amount of gas left.
+    /// @param _output_data  The pointer to the output.
+    /// @param _output_size  The output size.
+    result(evmc_status_code _status_code,
+           int64_t _gas_left,
+           const uint8_t* _output_data,
+           size_t _output_size) noexcept
+      : evmc_result{make_result(_status_code, _gas_left, _output_data, _output_size)}
+    {}
 
     /// Converting constructor from raw evmc_result.
     explicit result(evmc_result const& res) noexcept : evmc_result{res} {}
@@ -80,11 +341,32 @@ public:
 class vm
 {
 public:
+    vm() noexcept = default;
+
     /// Converting constructor from evmc_instance.
     explicit vm(evmc_instance* instance) noexcept : m_instance{instance} {}
 
     /// Destructor responsible for automatically destroying the VM instance.
-    ~vm() noexcept { m_instance->destroy(m_instance); }
+    ~vm() noexcept
+    {
+        if (m_instance)
+            m_instance->destroy(m_instance);
+    }
+
+    vm(const vm&) = delete;
+    vm& operator=(const vm&) = delete;
+
+    /// Move constructor.
+    vm(vm&& other) noexcept : m_instance{other.m_instance} { other.m_instance = nullptr; }
+
+    /// Move assignment operator.
+    vm& operator=(vm&& other) noexcept
+    {
+        this->~vm();
+        m_instance = other.m_instance;
+        other.m_instance = nullptr;
+        return *this;
+    }
 
     /// The constructor that captures a VM instance and configures the instance
     /// with provided list of options.
@@ -95,6 +377,9 @@ public:
         for (auto option : options)
             set_option(option.first, option.second);
     }
+
+    /// Checks if contains a valid pointer to the VM instance.
+    explicit operator bool() const noexcept { return m_instance != nullptr; }
 
     /// Checks whenever the VM instance is ABI compatible with the current EVMC API.
     bool is_abi_compatible() const noexcept { return m_instance->abi_version == EVMC_ABI_VERSION; }
@@ -128,7 +413,7 @@ public:
     }
 
 private:
-    evmc_instance* const m_instance = nullptr;
+    evmc_instance* m_instance = nullptr;
 };
 
 /// The EVMC Host interface
@@ -138,35 +423,33 @@ public:
     virtual ~HostInterface() noexcept = default;
 
     /// @copydoc evmc_host_interface::account_exists
-    virtual bool account_exists(const evmc_address& addr) noexcept = 0;
+    virtual bool account_exists(const address& addr) noexcept = 0;
 
     /// @copydoc evmc_host_interface::get_storage
-    virtual evmc_bytes32 get_storage(const evmc_address& addr,
-                                     const evmc_bytes32& key) noexcept = 0;
+    virtual bytes32 get_storage(const address& addr, const bytes32& key) noexcept = 0;
 
     /// @copydoc evmc_host_interface::set_storage
-    virtual evmc_storage_status set_storage(const evmc_address& addr,
-                                            const evmc_bytes32& key,
-                                            const evmc_bytes32& value) noexcept = 0;
+    virtual evmc_storage_status set_storage(const address& addr,
+                                            const bytes32& key,
+                                            const bytes32& value) noexcept = 0;
 
     /// @copydoc evmc_host_interface::get_balance
-    virtual evmc_uint256be get_balance(const evmc_address& addr) noexcept = 0;
+    virtual uint256be get_balance(const address& addr) noexcept = 0;
 
     /// @copydoc evmc_host_interface::get_code_size
-    virtual size_t get_code_size(const evmc_address& addr) noexcept = 0;
+    virtual size_t get_code_size(const address& addr) noexcept = 0;
 
     /// @copydoc evmc_host_interface::get_code_hash
-    virtual evmc_bytes32 get_code_hash(const evmc_address& addr) noexcept = 0;
+    virtual bytes32 get_code_hash(const address& addr) noexcept = 0;
 
     /// @copydoc evmc_host_interface::copy_code
-    virtual size_t copy_code(const evmc_address& addr,
+    virtual size_t copy_code(const address& addr,
                              size_t code_offset,
                              uint8_t* buffer_data,
                              size_t buffer_size) noexcept = 0;
 
     /// @copydoc evmc_host_interface::selfdestruct
-    virtual void selfdestruct(const evmc_address& addr,
-                              const evmc_address& beneficiary) noexcept = 0;
+    virtual void selfdestruct(const address& addr, const address& beneficiary) noexcept = 0;
 
     /// @copydoc evmc_host_interface::call
     virtual result call(const evmc_message& msg) noexcept = 0;
@@ -175,13 +458,13 @@ public:
     virtual evmc_tx_context get_tx_context() noexcept = 0;
 
     /// @copydoc evmc_host_interface::get_block_hash
-    virtual evmc_bytes32 get_block_hash(int64_t block_number) noexcept = 0;
+    virtual bytes32 get_block_hash(int64_t block_number) noexcept = 0;
 
     /// @copydoc evmc_host_interface::emit_log
-    virtual void emit_log(const evmc_address& addr,
+    virtual void emit_log(const address& addr,
                           const uint8_t* data,
                           size_t data_size,
-                          const evmc_bytes32 topics[],
+                          const bytes32 topics[],
                           size_t num_topics) noexcept = 0;
 };
 
@@ -198,39 +481,39 @@ public:
     /// Implicit converting constructor from evmc_context.
     HostContext(evmc_context* ctx) noexcept : context{ctx} {}  // NOLINT
 
-    bool account_exists(const evmc_address& address) noexcept final
+    bool account_exists(const address& address) noexcept final
     {
         return context->host->account_exists(context, &address);
     }
 
-    evmc_bytes32 get_storage(const evmc_address& address, const evmc_bytes32& key) noexcept final
+    bytes32 get_storage(const address& address, const bytes32& key) noexcept final
     {
         return context->host->get_storage(context, &address, &key);
     }
 
-    evmc_storage_status set_storage(const evmc_address& address,
-                                    const evmc_bytes32& key,
-                                    const evmc_bytes32& value) noexcept final
+    evmc_storage_status set_storage(const address& address,
+                                    const bytes32& key,
+                                    const bytes32& value) noexcept final
     {
         return context->host->set_storage(context, &address, &key, &value);
     }
 
-    evmc_uint256be get_balance(const evmc_address& address) noexcept final
+    uint256be get_balance(const address& address) noexcept final
     {
         return context->host->get_balance(context, &address);
     }
 
-    size_t get_code_size(const evmc_address& address) noexcept final
+    size_t get_code_size(const address& address) noexcept final
     {
         return context->host->get_code_size(context, &address);
     }
 
-    evmc_bytes32 get_code_hash(const evmc_address& address) noexcept final
+    bytes32 get_code_hash(const address& address) noexcept final
     {
         return context->host->get_code_hash(context, &address);
     }
 
-    size_t copy_code(const evmc_address& address,
+    size_t copy_code(const address& address,
                      size_t code_offset,
                      uint8_t* buffer_data,
                      size_t buffer_size) noexcept final
@@ -238,9 +521,9 @@ public:
         return context->host->copy_code(context, &address, code_offset, buffer_data, buffer_size);
     }
 
-    void selfdestruct(const evmc_address& address, const evmc_address& beneficiary) noexcept final
+    void selfdestruct(const address& addr, const address& beneficiary) noexcept final
     {
-        context->host->selfdestruct(context, &address, &beneficiary);
+        context->host->selfdestruct(context, &addr, &beneficiary);
     }
 
     result call(const evmc_message& message) noexcept final
@@ -261,18 +544,18 @@ public:
         return tx_context;
     }
 
-    evmc_bytes32 get_block_hash(int64_t number) noexcept final
+    bytes32 get_block_hash(int64_t number) noexcept final
     {
         return context->host->get_block_hash(context, number);
     }
 
-    void emit_log(const evmc_address& address,
+    void emit_log(const address& addr,
                   const uint8_t* data,
                   size_t data_size,
-                  const evmc_bytes32 topics[],
+                  const bytes32 topics[],
                   size_t topics_count) noexcept final
     {
-        context->host->emit_log(context, &address, data, data_size, topics, topics_count);
+        context->host->emit_log(context, &addr, data, data_size, topics, topics_count);
     }
 };
 
@@ -351,7 +634,8 @@ inline void emit_log(evmc_context* h,
                      const evmc_bytes32 topics[],
                      size_t num_topics) noexcept
 {
-    static_cast<Host*>(h)->emit_log(*addr, data, data_size, topics, num_topics);
+    static_cast<Host*>(h)->emit_log(*addr, data, data_size, static_cast<const bytes32*>(topics),
+                                    num_topics);
 }
 
 constexpr evmc_host_interface interface{
@@ -360,6 +644,42 @@ constexpr evmc_host_interface interface{
 };
 }  // namespace internal
 
-inline Host::Host() noexcept : evmc_context{&internal::interface} {}
+inline Host::Host() noexcept : evmc_context{&evmc::internal::interface} {}
 
 }  // namespace evmc
+
+
+namespace std
+{
+/// Hash operator template specialization for evmc::address. Needed for unordered containers.
+template <>
+struct hash<evmc::address>
+{
+    /// Hash operator using FNV1a-based folding.
+    constexpr size_t operator()(const evmc::address& s) const noexcept
+    {
+        using namespace evmc;
+        using namespace fnv;
+        return static_cast<size_t>(fnv1a_by64(
+            fnv1a_by64(fnv1a_by64(fnv::offset_basis, load64be(&s.bytes[0])), load64be(&s.bytes[8])),
+            load32be(&s.bytes[16])));
+    }
+};
+
+/// Hash operator template specialization for evmc::bytes32. Needed for unordered containers.
+template <>
+struct hash<evmc::bytes32>
+{
+    /// Hash operator using FNV1a-based folding.
+    constexpr size_t operator()(const evmc::bytes32& s) const noexcept
+    {
+        using namespace evmc;
+        using namespace fnv;
+        return static_cast<size_t>(
+            fnv1a_by64(fnv1a_by64(fnv1a_by64(fnv1a_by64(fnv::offset_basis, load64be(&s.bytes[0])),
+                                             load64be(&s.bytes[8])),
+                                  load64be(&s.bytes[16])),
+                       load64be(&s.bytes[24])));
+    }
+};
+}  // namespace std