+++ test/evmc/helpers.h
@@ -18,6 +18,8 @@
 #pragma once
 
 #include <evmc/evmc.h>
+#include <stdlib.h>
+#include <string.h>
 
 /**
  * Returns true if the VM instance has a compatible ABI version.
@@ -83,6 +85,7 @@ static inline enum evmc_set_option_result evmc_set_option(struct evmc_instance* 
  *
  * @see evmc_set_tracer_fn
  */
+EVMC_DEPRECATED
 static inline void evmc_set_tracer(struct evmc_instance* instance,
                                    evmc_trace_callback callback,
                                    struct evmc_tracer_context* context)
@@ -104,6 +107,58 @@ static inline struct evmc_result evmc_execute(struct evmc_instance* instance,
                                               size_t code_size)
 {
     return instance->execute(instance, context, rev, msg, code, code_size);
+}
+
+/// The evmc_result release function using free() for releasing the memory.
+///
+/// This function is used in the evmc_make_result(),
+/// but may be also used in other case if convenient.
+///
+/// @param result The result object.
+static void evmc_free_result_memory(const struct evmc_result* result)
+{
+    free((uint8_t*)result->output_data);
+}
+
+/// Creates the result from the provided arguments.
+///
+/// The provided output is copied to memory allocated with malloc()
+/// and the evmc_result::release function is set to one invoking free().
+///
+/// In case of memory allocation failure, the result has all fields zeroed
+/// and only evmc_result::status_code is set to ::EVMC_OUT_OF_MEMORY internal error.
+///
+/// @param status_code  The status code.
+/// @param gas_left     The amount of gas left.
+/// @param output_data  The pointer to the output.
+/// @param output_size  The output size.
+static inline struct evmc_result evmc_make_result(enum evmc_status_code status_code,
+                                                  int64_t gas_left,
+                                                  const uint8_t* output_data,
+                                                  size_t output_size)
+{
+    struct evmc_result result;
+    memset(&result, 0, sizeof(result));
+
+    if (output_size != 0)
+    {
+        uint8_t* buffer = (uint8_t*)malloc(output_size);
+
+        if (!buffer)
+        {
+            result.status_code = EVMC_OUT_OF_MEMORY;
+            return result;
+        }
+
+        memcpy(buffer, output_data, output_size);
+        result.output_data = buffer;
+        result.output_size = output_size;
+        result.release = evmc_free_result_memory;
+    }
+
+    result.status_code = status_code;
+    result.gas_left = gas_left;
+    return result;
 }
 
 /**